+++ src/REvoDesign/basic/__init__.py
@@ -1,4 +1,4 @@
-from .abc_singleton import SingletonAbstract
+from .abc_singleton import SingletonAbstract,reset_singletons
 from .abc_third_party_module import ThirdPartyModuleAbstract
 from .data_structure import IterableLoop
 from .designer import ExternalDesignerAbstract
@@ -11,6 +11,7 @@ from .server_monitor import MenuActionServerMonitor, ServerControlAbstract
 
 __all__ = [
     "SingletonAbstract",
+    "reset_singletons",
     "ThirdPartyModuleAbstract",
     "IterableLoop",
     "MutateRunnerAbstract",