+++ tests/conftest.py
@@ -23,20 +23,14 @@ from pytestqt import qtbot
 from RosettaPy.node import NodeHintT
 from RosettaPy.utils import tmpdir_manager
 
-from REvoDesign import ConfigBus, REvoDesignPlugin
+from REvoDesign import REvoDesignPlugin
+from REvoDesign.basic.abc_singleton import SingletonAbstract, reset_singletons
 from REvoDesign.bootstrap import EXPERIMENTS_CONFIG_DIR
-from REvoDesign.citations import CitationManager
-from REvoDesign.clients.QtSocketConnector import (REvoDesignWebSocketClient,
-                                                  REvoDesignWebSocketServer)
+
 from REvoDesign.common import MutantTree
-from REvoDesign.driver.file_dialog import FileDialog
-from REvoDesign.driver.ui_driver import StoresWidget
-from REvoDesign.editor.monaco.config import ConfigStore
-from REvoDesign.magician import Magician
-from REvoDesign.sidechain import SidechainSolver
+
 from REvoDesign.tools.customized_widgets import (get_widget_value,
                                                  set_widget_value)
-from REvoDesign.tools.system_tools import SystemInfoReduced
 
 from .data import TestData
 from .data.test_data import KeyData
@@ -84,13 +78,10 @@ def plugin(qtbot: qtbot.QtBot, app):
 
     cmd.reinitialize()
 
+    reset_singletons()
+
     # reset all singleton classes
-    CitationManager.reset_instance()
-    REvoDesignWebSocketClient.reset_instance()
-    REvoDesignWebSocketServer.reset_instance()
-    SidechainSolver.reset_instance()
-    StoresWidget.reset_instance()
-    ConfigBus.reset_instance()
+    
     gc.collect()
 
     plugin = REvoDesignPlugin()
@@ -504,19 +495,7 @@ class TestWorker:
         self.plugin.reinitialize()
         cmd.reinitialize()
 
-        # reset singleton classes
-        # in case of potential pollution of singleton instances accross tests,
-        # all subclasses inherit from SingletonAbstract must be reset immediately on teardown.
-        CitationManager.reset_instance()
-        REvoDesignWebSocketClient.reset_instance()
-        REvoDesignWebSocketServer.reset_instance()
-        SidechainSolver.reset_instance()
-        StoresWidget.reset_instance()
-        ConfigBus.reset_instance()
-        Magician.reset_instance()
-        FileDialog.reset_instance()
-        ConfigStore.reset_instance()
-        SystemInfoReduced.reset_instance()
+        reset_singletons()
 
         gc.collect()